Ability to clone a StringInput
This should be done for other types, but I'll just wait until I actually need
it to add it.

The purpose here is to be able to re-use a validation for a different field name.
There are plenty of places where this could be used, but the most obvious is
password & password_confirm fields which share the same validation.

It was tempting to define the validation separately from the field. So we just
end up with an array of validators that we can attach to any field. Except, as an
optimization, validators often contain error messages, which includes the field
name. So, to split the field from the validator, we'd have to create the error
messages at runtime. So it's a performance tradeoff.