Rework error responses so that error_id is in the body
Previously, the generated errorId was only written to the response's Error-Id
header. This has the advantage of allowing us to have fully static error
responses, but it's certainly less usable (e.g., callers need to log the body
AND a specific header). Adding an `error_id` field makes it easier for callers
to deal with errors.

Changed the log output to be a global variable. There's really no reason to have
a per instance output, and having it global gives us control when testing, which
is something the tests were struggling with.

The above two changes also resulted in how the http.Handler is implemented to
more consistently log requests in success and all error cases (previously,
loadEnv errors would only be consistent if the implementor made them so, now,
since we expect the response in return, we can force the consistency).